backend/src/requirements.txt
flask-sqlalchemy==2.5.1
marshmallow==3.15.0
hanlp_restful==0.0.17
mysqlclient==2.1.0
html2text==2020.1.16
requests==2.27.1
Flask-Limiter==2.4.5.1
pycryptodome==3.14.1
beautifulsoup4==4.9.0
lxml==4.8.0
